requirements.txt
Flask==3.0.3
feedparser==6.0.10
Jinja2==3.1.4
PyYAML==6.0.2
python-slugify==8.0.4
requests==2.32.3
Pillow==10.4.0
Markdown==3.6
